add Pattern#precludes?
this will be used for method insertion - rather than knowning about precision
and equivalence, I'm hoping to use this single notion of 'this pattern will
match in all cases that the other would'

for example, methods will be inserted at index N where N+1 is a pattern that
would preclude the one being added. if both patterns preclude each other,
however, the old pattern is replaced (this is what 'equivalence' was needed
for before).