remove #precludes?
the intent of #precludes? was to order method branch insertion based on
specificity of the pattern. this was convenient because it allowed
redefining of branches, with the new one taking precedence, and made the
order of method definitions not matter (e.g. n fib, 0 fib, 1 fib).

unfortunately it's very tricky to implement, the full scope of
preclude-detection being a combinatorial explosion (e.g. comparing And
with And, or And with *, or * with And), and i'm not sure the original
intent is such a good idea anymore: you probably shouldn't be redefining
methods all the time, and the order of precedence is much clearer if
simply laid out in code.

so, removing it. much simpler this way. we'll see how long it lasts.