perform pattern assignment at compile-time
this formalizes the pattern split into compile-time and run-time,
allowing us to do things like set local variables directly rather than
mutating VariableScopes at runtime.

should be way faster.

the code-side of patterns now provides an #assign(gen) methods, which is
called with the pattern and value on the stack, and is expected to
perform any assignment necessary and leave the stack as-is.